sign-language-predictor/api/requirements.txt
tensorflow-cpu==2.15.0
mediapipe==0.10.9
opencv-python-headless==4.9.0.80
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1.post1
joblib==1.3.2
flask==3.0.2
flask-cors==4.0.0
